rename: BinCallback.h -> .hpp
